homeassistant/addons/roamcore-victron-auto/requirements.txt
paho-mqtt==2.1.0
zeroconf==0.132.2
